requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
paho-mqtt==1.6.1
joblib==1.3.2
numpy==1.26.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
